Start making sorting the posts an option
With these changes, you can pass an sort function into the butler.

It would be nice to be able to give buter.getPosts a sort function, but maybe
not very necessary?

The index_manager.js file is not as pretty as I would like it to be. I'm
wondering if it can be cleaned up a bit...